+++ src/main/java/net/geforcemods/securitycraft/api/ICodebreakable.java
@@ -5,16 +5,25 @@ import net.minecraft.entity.player.EntityPlayer;
 
 /**
  * Marks a block as being able to be hacked with the Codebreaker.
- * 
+ *
  * @author Geforce
  */
 public interface ICodebreakable {
 	/**
-	 * Called when a Codebreaker is used on a block.
+	 * Checked before any codebreaking attempt, whether the codebreaker should attempt to break the code. Useful when the block
+	 * currently does not accept a code at all.
+	 *
+	 * @param state The state of the block that the codebreaking attempt should be performed on
+	 * @param player The player trying the codebreaking attempt
+	 * @return true if the codebreaking attempt should be performed, false otherwise
+	 */
+	public boolean shouldAttemptCodebreak(IBlockState state, EntityPlayer player);
+
+	/**
+	 * Called when a Codebreaker has successfully broken the code of a block.
 	 *
 	 * @param state The block state of the block.
 	 * @param player The player who used the Codebreaker.
-	 * @return Return true if the Codebreaker "hack" was successful, false otherwise.
 	 */
-	public boolean onCodebreakerUsed(IBlockState state, EntityPlayer player);
+	public void useCodebreaker(IBlockState state, EntityPlayer player);
 }