+++ src/main/java/net/geforcemods/securitycraft/blocks/BlockKeypad.java
@@ -63,7 +63,7 @@ public class BlockKeypad extends BlockDisguisable {
 				return true;
 			}
 			else if (!PlayerUtils.isHoldingItem(player, SCContent.codebreaker, hand))
-				te.openPasswordGUI(player);
+				te.openPasswordGUI(world, pos, te.getOwner(), player);
 		}
 
 		return true;