+++ src/main/java/net/geforcemods/securitycraft/blocks/BlockKeypadChest.java
@@ -128,7 +128,7 @@ public class BlockKeypadChest extends BlockOwnable {
 				activate(world, pos, player);
 			}
 			else if (!PlayerUtils.isHoldingItem(player, SCContent.codebreaker, hand))
-				te.openPasswordGUI(player);
+				te.openPasswordGUI(world, pos, te.getOwner(), player);
 		}
 
 		return true;