+++ src/main/java/net/geforcemods/securitycraft/blocks/BlockKeypadDoor.java
@@ -44,7 +44,7 @@ public class BlockKeypadDoor extends BlockSpecialDoor {
 				return true;
 			}
 			else if (!PlayerUtils.isHoldingItem(player, SCContent.codebreaker, hand))
-				te.openPasswordGUI(player);
+				te.openPasswordGUI(world, pos, te.getOwner(), player);
 		}
 
 		return true;